requirements.txt
cdbcrypt==3.1.4
cffi==1.11.5
Django==1.9.12
pycparser==2.18
six==1.11.0
